replace simple and numeric schemas with string, int, and float schemas

+++ pkg/services/mysql/all_in_one_provision_test.go
@@ -163,7 +163,7 @@ func TestValidateHardwareVersionIncompatible(t *testing.T) {
 	provisionSchema := planSchema{
 		allowedHardware:         []string{"gen5"},
 		defaultHardware:         "gen5",
-		allowedCores:            []int{2, 4, 8, 16},
+		allowedCores:            []int64{2, 4, 8, 16},
 		defaultCores:            2,
 		maxStorage:              2048,
 		minStorage:              5,
